backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
passlib[bcrypt]==1.7.4
cryptography==43.0.0
websockets==13.0
httpx==0.27.0
hyperliquid-python-sdk>=0.4.0
python-multipart==0.0.9
